[LLP] - added inheritance to LatLng animators
