+++ plugin-locationlayer/src/main/java/com/mapbox/mapboxsdk/plugins/locationlayer/LatLngEvaluator.java
@@ -1,0 +1,19 @@
+package com.mapbox.mapboxsdk.plugins.locationlayer;
+
+import android.animation.TypeEvaluator;
+
+import com.mapbox.mapboxsdk.geometry.LatLng;
+
+class LatLngEvaluator implements TypeEvaluator<LatLng> {
+
+  private final LatLng latLng = new LatLng();
+
+  @Override
+  public LatLng evaluate(float fraction, LatLng startValue, LatLng endValue) {
+    latLng.setLatitude(startValue.getLatitude()
+      + ((endValue.getLatitude() - startValue.getLatitude()) * fraction));
+    latLng.setLongitude(startValue.getLongitude()
+      + ((endValue.getLongitude() - startValue.getLongitude()) * fraction));
+    return latLng;
+  }
+}
